1.2.75 - Changes to how addon checks for updates.
Added some new triggers for when the addon checks for a new version, to get filter updates out to people quicker. It'll now check for a new version when you tab out of your game, and when the weather changes. (I know, I know, but I'm bad at this and don't really know how to get it to check on a timer or anything, so just using Windower events to do it instead, PRs welcome if anyone wants to make it less cursed)