1.4.45
More bug fixes, local testing shows it as working now. To make sure this never happens again, I've also added a new command, //sbl forceupdate. This will grab the current version from this repo, regardless of what version it is (or what version the lua thinks it is). I am beyond sorry for the trouble.